Structure as an installable package
