+++ pyproject.toml
@@ -1,0 +1,43 @@
+[build-system]
+requires = ["setuptools>=61.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "couch-remote"
+version = "0.1.0"
+description = "A simple remote web keyboard"
+readme = "README.md"
+authors = [{ name = "Sorebit", email = "tomulewicz.s@pm.me" }]
+license = { file = "LICENSE" }
+classifiers = [
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+    "Development Status :: 3 - Alpha",
+    "Environment :: Console",
+    "Environment :: Web Environment",
+    "Framework :: FastAPI",
+    "Intended Audience :: End Users/Desktop",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
+    "Topic :: Home Automation",
+    "Topic :: Utilities",
+]
+keywords = ["remote", "keyboard", "couch"]
+dependencies = [
+    "FastAPI>=0.79.0",
+    "uvicorn[standard]>=0.18.0",
+    "pydantic>=1.9.0,<2.0.0",
+    "jinja2>=3.1.2",
+    "pynput>=1.7.6",
+]
+requires-python = ">=3.8"
+
+[project.optional-dependencies]
+dev = ["black", "isort", "pip-tools", "pytest"]
+
+[project.urls]
+Homepage = "https://github.com/sorebit/pilot"
+
+[project.scripts]
+remote = "remote.__main__:main"
