+++ src/remote/statics/buttons.js
