+++ src/remote/statics/style.css
